dm: core: Walk the tree to find address- and size-cells properties
Walk the tree when reading size-cells or address-cells properties.

Reported-by: Robin Randhawa <Robin.Randhawa@ARM.com>
Signed-off-by: Matthias Brugger <mbrugger@suse.com>

Series-to: matthias.bgg@kernel.org
Series-to: sjg
Series-to: u-boot
Series-to: Robin.Randhawa@ARM.com
Series-cc: giulio.benetti@benettiengineering.com
Series-cc: treding@nvidia.com

Cover-letter:
Reading size-cells and address-cells from a node should walk up the
parent stack to find the property. This is reflected through comments in
the code, but actually not implemented. This leads to the fact that some
DTBs worked by accident. Fix this by walking the tree when reading this
properties.

After fixing the default size-cells returned when the property is not
present this problem was made visible. When adding flash devices in qemu
for arm64, the size-cells was wrongly inpterpreted and broke access to the
second flash bank.
END